Move `method_url` to separate file
